Pagination & DTO for pagination to work with our rest controller

+++ src/main/java/electro/by/gecko/vitrine/service/product/ProductDAOService.java
@@ -1,13 +1,11 @@
 package electro.by.gecko.vitrine.service.product;
 
 import electro.by.gecko.vitrine.entity.Product;
-import electro.by.gecko.vitrine.exception.ProductNotFoundException;
 import electro.by.gecko.vitrine.repository.ProductDAO;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.EntityNotFoundException;
 
 @Service
 public class ProductDAOService implements ProductService{
@@ -18,8 +16,8 @@ public class ProductDAOService implements ProductService{
     }
 
     @Override
-    public Product findById(Long id) throws EntityNotFoundException {
-        return productDAO.findById(id).orElseThrow(() -> new ProductNotFoundException(id));
+    public Product findById(Long id) {
+        return productDAO.findById(id).orElse(null); //Throw(() -> new ProductNotFoundException(id));
     }
 
     @Override
@@ -28,9 +26,9 @@ public class ProductDAOService implements ProductService{
     }
 
     @Override
-    public void deleteById(Long id) throws EntityNotFoundException {
-        productDAO.findById(id).orElseThrow(() -> new ProductNotFoundException(id));
-        productDAO.deleteById(id);
+    public void deleteById(Long id) {
+        if(null != productDAO.findById(id).orElse(null))  //.orElseThrow(() -> new ProductNotFoundException(id));
+            productDAO.deleteById(id);
     }
 
     @Override
@@ -38,12 +36,8 @@ public class ProductDAOService implements ProductService{
         return productDAO.findAll();
     }
 
-    public Page<Product> findPage(Pageable page) {
-        return productDAO.findAll(page);
-    }
-
     @Override
-    public Iterable<Product> findByFilter(String name, String description, int minPrice, int maxPrice, String brand, Pageable page) {
-        return productDAO.findFilter(name, description, minPrice, maxPrice, brand, page);
+    public Page<Product> findByFilter(String search, int minPrice, int maxPrice, String brand, Pageable page) {
+        return productDAO.findFilter(search, minPrice, maxPrice, brand, page);
     }
 }
